docker_gpio_i2c/requirements.txt
smbus2
RPi.bme280
